refactor(mobile): improve Receive routing
